Refactor data handling for LOSOCV training
- Updated .gitignore to exclude subject data directory.
- Modified train_colab.ipynb to implement a per-subject caching mechanism for preprocessing, reducing RAM usage during training.
- Added a new data_pipeline.py module to manage disk caching and tf.data streaming for training data.
- Changed data types in data_loader.py from float64 to float32 for EEG/ECG stimuli to optimize memory usage.
- Refactored losocv_train function in train.py to utilize the new data pipeline, allowing for efficient data loading and augmentation.
- Ensured that the training process streams data from disk, significantly lowering peak RAM requirements.

+++ src/data_loader.py
@@ -52,10 +52,10 @@ def extract_subject_data(dreamer, subject_idx):
         eeg_base = subject["EEG"][0, 0]["baseline"][0, 0][trial_idx, 0]
         ecg_base = subject["ECG"][0, 0]["baseline"][0, 0][trial_idx, 0]
 
-        eeg_stimuli.append(eeg_stim.astype(np.float64))
-        ecg_stimuli.append(ecg_stim.astype(np.float64))
-        eeg_baseline.append(eeg_base.astype(np.float64))
-        ecg_baseline.append(ecg_base.astype(np.float64))
+        eeg_stimuli.append(eeg_stim.astype(np.float32))
+        ecg_stimuli.append(ecg_stim.astype(np.float32))
+        eeg_baseline.append(eeg_base.astype(np.float32))
+        ecg_baseline.append(ecg_base.astype(np.float32))
 
     valence = subject["ScoreValence"][0, 0].flatten().astype(int)
     arousal = subject["ScoreArousal"][0, 0].flatten().astype(int)
